This is a compatibility breaking change! * Update gem naming to follow convention   * Note this is useful because it prevents confusion for users and will     avoid a class of naming collision issues users may run into
